replace materials api code from mvc project to materials api project
