+++ Frontend/SciMaterials.UI.MVC/Views/_ViewImports.cshtml
@@ -1,5 +1,4 @@
 ﻿@using SciMaterials.UI.MVC
-@using SciMaterials.UI.MVC.API
 @using SciMaterials.UI.MVC.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
