+++ Materials/SciMaterials.Materials.Api/Controllers/CommentsController.cs
@@ -1,11 +1,10 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using SciMaterials.Contracts.API.Constants;
-using SciMaterials.Contracts.API.DTO.Authors;
 using SciMaterials.Contracts.API.DTO.Comments;
 using SciMaterials.Contracts.API.Services.Comments;
 using SciMaterials.Contracts.Result;
 
-namespace SciMaterials.UI.MVC.API.Controllers;
+namespace SciMaterials.Materials.Api.Controllers;
 
 /// <summary> Service for working with comments. </summary>
 [ApiController]
