+++ Materials/SciMaterials.Materials.Api/Controllers/TagsController.cs
@@ -1,71 +1,67 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using SciMaterials.Contracts.API.Constants;
-using SciMaterials.Contracts.API.DTO.Authors;
-using SciMaterials.Contracts.API.DTO.Categories;
-using SciMaterials.Contracts.API.Services.Authors;
+using SciMaterials.Contracts.API.DTO.Tags;
+using SciMaterials.Contracts.API.Services.Tags;
 using SciMaterials.Contracts.Result;
-using SciMaterials.Contracts.ShortLinks;
-using SciMaterials.Services.ShortLinks;
 
-namespace SciMaterials.UI.MVC.API.Controllers;
+namespace SciMaterials.Materials.Api.Controllers;
 
 /// <summary> Service for working with authors. </summary>
 [ApiController]
-[Route(WebApiRoute.Authors)]
-public class AuthorsController : ApiBaseController<AuthorsController>
+[Route(WebApiRoute.Tags)]
+public class TagsController : ApiBaseController<TagsController>
 {
-    private readonly IAuthorService _authorService;
+    private readonly ITagService _authorService;
 
-    public AuthorsController(IAuthorService authorService, ILinkReplaceService link)
+    public TagsController(ITagService authorService)
     {
         _authorService = authorService;
     }
 
-    /// <summary> Get All Authors. </summary>
+    /// <summary> Get All Tags. </summary>
     /// <returns> Status 200 OK. </returns>
     [HttpGet]
-    [ProducesDefaultResponseType(typeof(Result<IEnumerable<GetAuthorResponse>>))]
+    [ProducesDefaultResponseType(typeof(Result<IEnumerable<GetTagResponse>>))]
     public async Task<IActionResult> GetAllAsync()
     {
-        _logger.LogInformation("Get all");
         var сategories = await _authorService.GetAllAsync();
         return Ok(сategories);
     }
 
-    /// <summary> Get Author by Id. </summary>
-    /// <param name="id"> Author Id. </param>
+    /// <summary> Get Tag by Id. </summary>
+    /// <param name="id"> Tag Id. </param>
     [HttpGet("{id}")]
-    [ProducesDefaultResponseType(typeof(Result<GetAuthorResponse>))]
+    [ProducesDefaultResponseType(typeof(Result<GetTagResponse>))]
     public async Task<IActionResult> GetByIdAsync([FromRoute] Guid id)
     {
         var products = await _authorService.GetByIdAsync(id);
         return Ok(products);
     }
 
-    /// <summary> Add a Author. </summary>
+    /// <summary> Add a Tag. </summary>
     /// <param name="request"> Add Request DTO. </param>
     /// <returns> Status 200 OK. </returns>
     [HttpPost("Add")]
     [ProducesDefaultResponseType(typeof(Guid))]
-    public async Task<IActionResult> AddAsync([FromBody] AddAuthorRequest request)
+    public async Task<IActionResult> AddAsync([FromBody] AddTagRequest request)
     {
         var result = await _authorService.AddAsync(request);
         return Ok(result);
     }
 
-    /// <summary> Edit a Author. </summary>
+    /// <summary> Edit a Tag. </summary>
     /// <param name="request"> Edit Request DTO. </param>
     /// <returns> Status 200 OK. </returns>
     [HttpPut("Edit")]
     [ProducesDefaultResponseType(typeof(Guid))]
-    public async Task<IActionResult> EditAsync([FromBody] EditAuthorRequest request)
+    public async Task<IActionResult> EditAsync([FromBody] EditTagRequest request)
     {
         var result = await _authorService.EditAsync(request);
         return Ok(result);
     }
 
-    /// <summary> Delete a Author. </summary>
-    /// <param name="id"> Author Id. </param>
+    /// <summary> Delete a Tag. </summary>
+    /// <param name="id"> Tag Id. </param>
     /// <returns> Status 200 OK response. </returns>
     [HttpDelete("{id}")]
     [ProducesDefaultResponseType(typeof(Guid))]
