+++ Materials/SciMaterials.Materials.Api/Filters/ReApplyOptionalRouteParameterOperationFilter.cs
@@ -1,8 +1,7 @@
 using Microsoft.OpenApi.Models;
-
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace SciMaterials.UI.MVC.API.Filters;
+namespace SciMaterials.Materials.Api.Filters;
 public class ReApplyOptionalRouteParameterOperationFilter : IOperationFilter
 {
 	const string captureName = "routeParameter";
